fix: security bugfixes for 7.14.2
Bug 1 (TRON display mismatch): fsm_msg_tron.h displayed deprecated
to_address/amount fields that are NOT part of raw_data being signed.
A malicious host could show one transfer and get a different tx signed.
Replaced with a single blind-sign prompt showing only the raw_data size.

Bug 5 (ETH RLP leading zeros): ethereum.c hashed nonce/gas/value via
hash_rlp_field which preserves leading zero bytes. Per the Ethereum yellow
paper, integer fields must have leading zeros stripped. Added
hash_rlp_bytes_stripped() and applied it to nonce, gas_price, gas_limit,
max_fee_per_gas, max_priority_fee_per_gas, and value.

Bug 6 (EIP-712 cancel ignored): confirmName() and confirmValue() cast
review() return value to void, allowing signing to continue after user
cancels. Added USER_CANCELLED error code and propagate cancellation.

Bug 7 (EIP-712 strtoll overflow): strtoll silently wraps values above
int64_max. Added errno check, endptr validation, and rejection of
negative values for uint types.

Bug 8 (token chain_id truncation): TokenType.chain_id was uint8_t,
silently truncating chainIds > 255 (Base=8453, Arbitrum=42161, etc.).
Widened to uint32_t throughout: struct, tokenByChainAddress(),
tokenByTicker().

+++ include/keepkey/firmware/ethereum_tokens.h
@@ -39,7 +39,7 @@ enum {
 typedef struct _TokenType {
   const char* const address;
   const char* const ticker;
-  uint8_t chain_id;
+  uint32_t chain_id;
   uint8_t decimals;
 } TokenType;
 
@@ -51,7 +51,7 @@ extern const TokenType* UnknownToken;
 
 const TokenType* tokenIter(int32_t* ctr);
 
-const TokenType* tokenByChainAddress(uint8_t chain_id, const uint8_t* address);
+const TokenType* tokenByChainAddress(uint32_t chain_id, const uint8_t* address);
 
 /// Tokens don't have unique tickers, so this might not return the one you're
 /// looking for :/
@@ -64,7 +64,7 @@ const TokenType* tokenByChainAddress(uint8_t chain_id, const uint8_t* address);
 /// \param[out] token The found token, assuming it was uniquely determinable.
 /// \returns true iff the token can be uniquely found in the list of known
 /// tokens.
-bool tokenByTicker(uint8_t chain_id, const char* ticker,
+bool tokenByTicker(uint32_t chain_id, const char* ticker,
                    const TokenType** token);
 
 void coinFromToken(CoinType* coin, const TokenType* token);
